% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fweibull.R
\name{fweibull}
\alias{fweibull}
\title{Weibull vulnerability curve}
\usage{
fweibull(P, SX, PX, X = 50)
}
\arguments{
\item{P}{Water potential (positive-valued MPa)}

\item{SX}{Shape parameter}

\item{PX}{Water potential at X loss of conductivity.}

\item{X}{If 50, PX is the P50.}
}
\description{
As re-parameterized by Ogle et al.
}

