% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitplc-package.R
\docType{data}
\name{stemvul}
\alias{stemvul}
\title{An example vulnerability curve}
\format{\describe{
\item{Species}{One of dpap, egran, ssay}
\item{Branch}{Replicate branch, multiple branches were measured for each species}
\item{MPa}{Xylem water potential (MPa)}
\item{PLC}{Percent loss conductivity}
\item{Cond}{Raw, unscaled conductivity of branch segment (units)}
}}
\description{
Percent loss conductivity as a function of water potential for three species.
}
