% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPx.R
\name{getPx}
\alias{getPx}
\title{Extract Px from fitted objects}
\usage{
getPx(object, x = 50)
}
\arguments{
\item{object}{Object returned by \code{\link{fitplc}}}

\item{x}{The x in Px, that is, if P50 should be returned, x=50.}
}
\description{
Extract esimates of Px from an object returned by \code{\link{fitplc}}. This allows extraction of estimates of P88 or other values when the fit estimated P50 (or other), for example. When the Weibull model is used, it is especially recommended to fit the P50 and estimate other points of the curve with \code{getPx}. 

See examples for use of this function. Note that the confidence interval is based on the bootstrap resampling performed by \code{\link{fitplc}}. This function only works when \code{bootci=TRUE} when the curve was fit.
}
\details{
Note that this function does not return a standard error, because the bootstrap confidence interval will be rarely symmetrical. If you like, you can calculate it as the mean of the half CI width (and note it as an 'approximate standard error'). Or, perhaps better, just report the CI and not the SE.

Also, frequently only the lower CI will be reported - sometimes the upper CI cannot be calculated (this will be more common when x is large, say for P88). In that case, assume symmetry and construct the CI with the lower confidence limit that will be reported. 

Finally, sometimes when x=88, it will return a missing value also for the predicted parameter. This happens when the predictions of the fitted model within the range of the data don't reach 88% embolism. We should probably avoid extrapolating the fit.
}
\examples{
\dontrun{
# Make sure bootci=TRUE (the default)
somefit <- fitplc(dfr, x=50)

# Extract P12
getPx(somefit, x=12)

}
}

